requirements.txt
streamlit
langchain-community
langchain-core
langchain-huggingface
sentence-transformers
faiss-cpu
transformers
huggingface_hub
pymupdf
markdown
